Moved qna related classes under qna namespace + style changes
