+++ DTML.EduBot/Dialogs/RootDialog.cs
@@ -1,14 +1,13 @@
-﻿using System;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.Bot.Builder.Dialogs;
-using Microsoft.Bot.Builder.Luis;
-using Microsoft.Bot.Builder.Luis.Models;
-using DTML.EduBot.Common;
-using DTML.EduBot.Services;
-
-namespace DTML.EduBot.Dialogs
+﻿namespace DTML.EduBot.Dialogs
 {
+    using System;
+    using System.Threading.Tasks;
+    using Microsoft.Bot.Builder.Dialogs;
+    using Microsoft.Bot.Builder.Luis;
+    using Microsoft.Bot.Builder.Luis.Models;
+    using DTML.EduBot.Common;
+    using DTML.EduBot.Qna;
+
     [LuisModel("", "")]
     [QnaModel(hostUri: "", subscriptionKey: "", modelId: "")]
     [Serializable]
