+++ DTML.EduBot/Extensions/Set.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-
-namespace DTML.EduBot.Extensions
+﻿namespace DTML.EduBot.Extensions
 {
     public static class Set
     {
@@ -11,17 +6,6 @@ namespace DTML.EduBot.Extensions
         {
             argument.NotNull(argumentName);
             target = argument;
-        }
-    }
-
-    public static class Require
-    {
-        public static void NotNull(this object argument, string argumentName)
-        {
-            if (argument == null)
-            {
-                throw new ArgumentNullException(argumentName);
-            }
         }
     }
 }
