+++ DTML.EduBot/Qna/IQnaResult.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace DTML.EduBot.Services
+namespace DTML.EduBot.Qna
 {
     public interface IQnaResult
     {
