+++ DTML.EduBot/Qna/QnaResult.cs
@@ -1,10 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-
-namespace DTML.EduBot.Services
+﻿namespace DTML.EduBot.Qna
 {
+    using System;
+    using System.Collections.Generic;
+
     [Serializable]
     public class QnaResult : IQnaResult
     {
