+++ DTML.EduBot/Qna/QnaResultCollection.cs
@@ -1,0 +1,7 @@
+﻿namespace DTML.EduBot.Qna
+{
+    public class QnaResultCollection
+    {
+        public QnaResult[] Answers { get; set; }
+    }
+}
