+++ DTML.EduBot/Qna/QnaService.cs
@@ -1,7 +1,6 @@
-﻿namespace DTML.EduBot.Services
+﻿namespace DTML.EduBot.Qna
 {
     using System;
-    using System.Collections.Generic;
     using System.Linq;
     using System.Net.Http;
     using System.Text;
