--- DTML.EduBot/Services/QnaResultCollection.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-
-namespace DTML.EduBot.Services
-{
-    public class QnaResultCollection
-    {
-        public QnaResult[] Answers { get; set; }
-    }
-}
